requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
autopep8==1.5.4
certifi==2020.12.5
Django @ file:///tmp/build/80754af9/django_1609784082548/work
mysql-connector==2.2.9
mysql-connector-python==8.0.23
numpy==1.19.5
pandas==1.2.1
protobuf==3.14.0
psycopg2 @ file:///C:/ci/psycopg2_1608147681824/work
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml==0.10.2
wincertstore==0.2
